DEV : access database for info on trains
create a database access layer, which will contain basic info about
the train for now. the database is accessed to get the origin and
destination information

NOTE : the database entries were created by hand for testing. There
needs to be an automated way to create the database entries from the
existing CSV data

	modified:   trains/__init__.py
	new file:   trains/db.py
	new file:   trains/schema.sql